requirements.txt
tldextract
dnsgen
